wip refactoring bridge utils
