+++ src/utils/bridge/api.ts
@@ -1,0 +1,3 @@
+import { api } from '@sora-substrate/util';
+
+export const bridgeApi = api.bridge;
